requirements.txt
torch==2.2.0
torchvision==0.17.0
Pillow==10.2.0
pandas==2.2.0
pyarrow==15.0.0
opencv-python == 4.9.0.80
matplotlib==3.8.2